Split `MoraleIndex` enum from `Common.h`
Note: The enum name is actually `MoraleIndexs`. This is a typo. The plural is also inconsistent with other enum names.
