+++ OPHD/EnumMoraleIndex.h
@@ -1,0 +1,23 @@
+#pragma once
+
+
+enum class MoraleIndexs
+{
+	Terrible,
+	Poor,
+	Fair,
+	Good,
+	Excellent,
+
+	Description,
+	Births,
+	Deaths,
+	NoFoodProduction,
+	Parks,
+	Recreation,
+	Commercial,
+	ResidentialOverflow,
+	BiowasteOverflow,
+	StructuresDisabled,
+	StructuresDestroyed
+};
